Added house, added mousewheel
Mouse wheel basically works as boost.. you still need to face the way
you want to zoom in/out (just moves much faster towards that direction).
Texturing of an obj sucks! I thought we could grab separate objects from
the obj file, but apparently we cant, unless we want to rewrite the
whole ajax code (which we wont). Another nicer way of making a house is
adding more sqaures, so the texturing is applied to each sqaure, instead
of stretched weirdly over the whole wall.  Also alternative would be
making just a separate HouseNode and creating there all the squares. The
hardest would be to make a roof.